requirements.txt
streamlit
pandas 
numpy 
tldextract 
plotly 
scikit-learn 
python-whois 
requests